services/product_service/requirements.txt
django
djangorestframework
django-filter
psycopg2-binary
python-dotenv
django-redis
redis
django-cors-headers
requests
